add function to check contrast
